task 5 almost done

+++ ConsoleApplication1/HitInfo.cpp
@@ -3,5 +3,5 @@
 
 
 HitInfo::HitInfo() { }
-HitInfo::HitInfo(Float3 point, Float3 normal, float distance) : p(point), n(normal), d(distance) { }
+HitInfo::HitInfo(Float3 point, Float3 normal, double distance) : p(point), n(normal), d(distance) { }
 HitInfo::~HitInfo() { }
